pyproject.toml
[tool.poetry]
name = "github-repo-loc-analyser"
version = "0.1.0"
description = "Software to analyse the LOC and LOCC for random github repos"
authors = ["Tim Neumann <neumantm@fius.informatik.uni-stuttgart.de>"]
license = "MIT"

[tool.poetry.scripts]
grla = 'github_repo_loc_analyser.main:main'

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
celery = "^5.0.5"
GitPython = "^3.1.12"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.1.1"
flake8-docstrings = "^1.5.0"
flake8-bugbear = "^20.1.4"

[tool.black]
line-length = 120
include = '\.pyi?$'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
